requirements.txt
Flask >= 2.1
pyyaml >= 5.3